+++ src/app/(auth)/sign-in/_components/anonymous-sign-in-button.tsx
